rename model-level datalibs
