+++ docs/api.rst
@@ -55,11 +55,11 @@ ECMWF
     :toctree: api/
 
     datalib.ecmwf.ERA5
+    datalib.ecmwf.era5_model_level
     datalib.ecmwf.HRES
+    datalib.ecmwf.hres_model_level
     datalib.ecmwf.hres.get_forecast_filename
     datalib.ecmwf.model_levels
-    datalib.ecmwf.model_level_era5
-    datalib.ecmwf.model_level_hres
     datalib.ecmwf.IFS
     datalib.ecmwf.variables
